+++ packages/shared/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./access";
 export * from "./domain";
 export * from "./dates";
 export * from "./date-utils";